stop using size_t, use unsigned int instead

+++ cpurast/canvas.hpp
@@ -4,7 +4,6 @@
 
 #include "color.hpp"
 #include <vector>
-#include <cstddef>
 
 namespace cr
 {
@@ -15,6 +14,6 @@ namespace cr
         virtual ~canvas() = default;
 
         // canvas resolution and color buffer resolution must be the same
-        virtual void draw(const std::vector<color>& color_buf, size_t width, size_t height) = 0;
+        virtual void draw(const std::vector<color>& color_buf, unsigned int width, unsigned int height) = 0;
     };
 }
