+++ cpurast/viewport.hpp
@@ -2,15 +2,13 @@
 
 #pragma once
 
-#include <cstddef>
-
 namespace cr
 {
     struct viewport
     {
-        size_t x;
-        size_t y;
-        size_t width;
-        size_t height;
+        unsigned int x;
+        unsigned int y;
+        unsigned int width;
+        unsigned int height;
     };
 }
